requirements.txt
Django==4.2.9
psycopg2-binary==2.9.9
redis==5.0.1
gunicorn==21.2.0
whitenoise==6.6.0
python-decouple==3.8
httpx==0.26.0
minio==7.2.3
requests==2.31.0
django-environ==0.11.2
django-allauth==0.61.1
django-crispy-forms==2.1
crispy-bootstrap5==0.7
Pillow==10.2.0
python-dateutil==2.8.2
PyPDF2==3.0.1
python-docx==1.1.0
markdown==3.5.1
django-formify==0.0.8
django-viewcomponent==1.0.10